Removed two model classes and update userManagment part

+++ src/components/userManagement-components/profile/employee-profile.component.js
@@ -13,7 +13,6 @@ export default class EmployeeProfile extends Component {
         this.employeeLogout = this.employeeLogout.bind(this);
 
         this.state = {
-            // Employees: [],
             empID: '',
             firstName: '',
             lastName: '',
@@ -49,19 +48,29 @@ export default class EmployeeProfile extends Component {
     }
 
     employeeLogout() {
-        this.props.history.push('/employee-registration');      
+        this.props.history.push('/employee-registration');
     }
 
     editEmployeeDetails() {
-        this.props.history.push(`/update-employee-details/${this.props.match.params.id}`);      
+        this.props.history.push(`/update-employee-details/${this.props.match.params.id}`);
     }
 
     deleteEmployeeDetails(id) {
-        
-        axios.delete('http://localhost:5000/employee/registration/' + this.props.match.params.id)
-            .then(res => console.log(res.data));
-           alert('Employee details are deleted.');
-           this.props.history.push('/employee-registration');
+
+        // axios.delete('http://localhost:5000/employee/registration/' + this.props.match.params.id)
+        //     .then(res => console.log(res.data));
+        // alert('Employee details are deleted.');
+        // this.props.history.push('/employee-registration');
+
+        const confirmBox = window.confirm(
+            "Are you sure want to delete your account?"
+        )
+        if (confirmBox === true) {
+            axios.delete('http://localhost:5000/employee/registration/' + this.props.match.params.id)
+                .then(res => console.log(res.data));
+            alert('Your account delete successfully!');
+            this.props.history.push('/employee-registration');
+        }
 
     }
 
@@ -69,13 +78,13 @@ export default class EmployeeProfile extends Component {
 
     render() {
         return (
-            <div style={{ marginLeft:"-200px"}}  className={styles.container}>
-                
+            <div style={{ marginLeft: "-200px" }} className={styles.container}>
+
                 <br />
                 <form className={styles.form_container} onSubmit={this.editEmployeeDetails}>
 
                     <div className="form-group">
-                    <h3>Profile</h3>
+                        <h3>Profile</h3>
                         <label>Profile photo:</label>
                         <div><img style={{ width: "200px", height: "200px" }} src={this.state.image}></img></div>
                     </div>
@@ -115,8 +124,8 @@ export default class EmployeeProfile extends Component {
                     <div className="form-group">
                         <input type="submit" value="Update" className="btn btn-primary" />
                         <button style={{ marginLeft: "10px" }} className="btn btn-danger" type='button' onClick={this.deleteEmployeeDetails}> Delete</button>
-                        <button style={{ marginLeft: "10px" }} className="btn btn-danger" type='button' onClick={this.employeeLogout}> LogOut</button>
-                        
+                        <button style={{ marginLeft: "10px" }} className="btn btn-danger" type='button' onClick={this.employeeLogout}> Logout</button>
+
                     </div>
                 </form>
